pyproject.toml
[project]
name = "wcrod"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "geopandas>=1.1.1",
    "polars>=1.31.0",
    "psycopg>=3.2.9",
    "sqlalchemy>=2.0.41",
]
